feat(handdecomp): port NewEfxFlashHPBar/Unit/UnitEffectEnd → matching C
screen NEAR 0c->14 = s16 sign-ext (lsr->asr); int local/param + (s16) cast
forces asr. matching-C 7642→7645.

+++ include/efxbattle.h
@@ -2697,7 +2697,7 @@ void EndEfxHPBarColorChange(void);
 void DisableEfxHpBarColorChange(void);
 void EnableEfxHpBarColorChange(void);
 void EfxHPBarColorChangeMain(struct ProcEfxHpBarColorChange * proc);
-void NewEfxFlashUnit(struct Anim * anim, u16 dura1, u16 dura2, int c);
+void NewEfxFlashUnit(struct Anim * anim, int dura1, int dura2, int c);
 void EfxFlashUnitMain(struct ProcEfxFlashing * proc);
 void EfxFlashUnitRestorePal(struct ProcEfxFlashing * proc);
 void NewEfxFlashUnitEffectEnd(struct Anim * anim, int dura1, int dura2, int c);
